requirements.txt
uvicorn~=0.34.3
fastapi~=0.115.9
pydantic~=2.11.7
langchain~=0.3.25
langchain-community~=0.3.25
langchain-openai~=0.3.23
langchain-core~=0.3.68
langgraph~=0.5.2
pinecone~=7.3.0
python-dotenv~=1.1.0